requirements.txt
fastapi==0.115.11
uvicorn==0.34.0

environ-config==24.1.0
python-dotenv==1.0.1